Added multi-role authentication

+++ app/Http/Controllers/RoomController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Room;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class RoomController extends Controller
 {
@@ -12,6 +13,9 @@ class RoomController extends Controller
      */
     public function index()
     {
+        if (Gate::denies('view', Room::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $rooms = Room::paginate(5);
         return view('admin.rooms.index', compact('rooms'));
     }
@@ -21,6 +25,9 @@ class RoomController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('create', Room::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         return view('admin.rooms.create');
     }
 
@@ -29,6 +36,9 @@ class RoomController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('create', Room::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $request->validate([
             'building' => 'required|string|max:10',
             'name' => 'required|string|max:10',
@@ -51,6 +61,9 @@ class RoomController extends Controller
      */
     public function edit(string $id)
     {
+        if (Gate::denies('update', Room::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $room = Room::find($id);
         return view('admin.rooms.edit', compact('room'));
     }
@@ -60,6 +73,9 @@ class RoomController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (Gate::denies('update', Room::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $request->validate([
             'building' => 'required|string|max:10',
             'name' => 'required|string|max:10',
@@ -76,8 +92,25 @@ class RoomController extends Controller
      */
     public function destroy(string $id)
     {
+        if (Gate::denies('delete', Room::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $room = Room::find($id);
         $room->delete();
         return redirect()->route('rooms.index')->with('success', 'Room deleted successfully.');
+    }
+
+    public function search(Request $request){
+        $searchData = $request->search_data;
+        if($searchData == ""){
+            return redirect()->route('rooms.index');
+        } else {
+            $rooms = Room::where('name','LIKE',"%".$searchData."%")
+            ->orWhere('id', '=',$searchData)
+            ->orWhere('building','LIKE','%'.$searchData.'%')
+            ->orWhere('name','LIKE','%'.$searchData.'%')
+            ->paginate(5);
+            return view('admin.rooms.index', compact('rooms'));
+        }
     }
 }
