+++ app/Http/Controllers/TeacherController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Teacher;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TeacherController extends Controller
 {
@@ -12,6 +13,9 @@ class TeacherController extends Controller
      */
     public function index()
     {
+        if (Gate::denies('view', Teacher::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $teachers = Teacher::paginate(5);
         return view('admin.teachers.index', compact('teachers'));
     }
@@ -21,6 +25,9 @@ class TeacherController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('create', Teacher::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         return view('admin.teachers.create');
     }
 
@@ -29,6 +36,9 @@ class TeacherController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('create', Teacher::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $request->validate([
             'name'=>'required|string|max:50',
             'subject'=>'required|string|max:50',
@@ -53,6 +63,9 @@ class TeacherController extends Controller
      */
     public function edit(string $id)
     {
+        if (Gate::denies('update', Teacher::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $teacher = Teacher::find($id);
         return view('admin.teachers.edit', compact('teacher'));
     }
@@ -62,6 +75,9 @@ class TeacherController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (Gate::denies('update', Teacher::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $request->validate([
             'name'=>'required|string|max:50',
             'subject'=>'required|string|max:50',
@@ -80,8 +96,26 @@ class TeacherController extends Controller
      */
     public function destroy(string $id)
     {
+        if (Gate::denies('delete', Teacher::class)) {
+            return redirect()->route("admin.dashboard")->with('error', 'No permission.');
+        }
         $teacher = Teacher::find($id);
         $teacher->delete();
         return redirect()->route('teachers.index')->with('success', 'Teacher deleted successfully.');
+    }
+
+    public function search(Request $request){
+        $searchData = $request->search_data;
+        if($searchData == ""){
+            return redirect()->route('teachers.index');
+        } else {
+            $teachers = Teacher::where('name','LIKE',"%".$searchData."%")
+            ->orWhere('id', '=',$searchData)
+            ->orWhere('name','LIKE','%'.$searchData.'%')
+            ->orWhere('email','LIKE','%'.$searchData.'%')
+            ->orWhere('subject','LIKE','%'.$searchData.'%')
+            ->paginate(5);
+            return view('admin.teachers.index', compact('teachers'));
+        }
     }
 }
